% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{basic_composition}
\alias{basic_composition}
\title{Basic (sequential) composition}
\usage{
basic_composition(epsilons, deltas = rep(0, length(epsilons)))
}
\arguments{
\item{epsilons}{Numeric vector of per-release epsilons.}

\item{deltas}{Numeric vector of per-release deltas.}
}
\value{
List with total \code{epsilon} and \code{delta}.
}
\description{
Basic (sequential) composition
}
