% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{DPrivStats-package}
\alias{DPrivStats}
\alias{DPrivStats-package}
\title{DPrivStats: Differentially Private Classical Inference}
\description{
Implements differentially private versions of common statistical
procedures with privacy budget accounting and privacy-aware confidence
intervals.
}
\author{
\strong{Maintainer}: Mukul Bijalwan \email{mukulbijalwan555@gmail.com}

Authors:
\itemize{
  \item Mukul Bijalwan \email{mukulbijalwan555@gmail.com}
}

}
\keyword{internal}
