% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{epsilon_to_rdp}
\alias{epsilon_to_rdp}
\title{Convert epsilon-DP to zCDP (RDP) parameter}
\usage{
epsilon_to_rdp(epsilon)
}
\arguments{
\item{epsilon}{Pure-DP epsilon.}
}
\value{
zCDP rho parameter.
}
\description{
An (epsilon, 0)-DP mechanism satisfies (epsilon^2 / 2)-zCDP
(Bun & Steinke, 2016).
}
