% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.dp_confint}
\alias{print.dp_confint}
\title{Print method for dp_confint objects}
\usage{
\method{print}{dp_confint}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{dp_confint} matrix.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments (ignored).}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print method for dp_confint objects
}
