% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.dp_estimate}
\alias{print.dp_estimate}
\title{Print method for dp_estimate objects}
\usage{
\method{print}{dp_estimate}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{dp_estimate} object.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments (ignored).}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print method for dp_estimate objects
}
