% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_regression2.R
\name{print.dp_glm}
\alias{print.dp_glm}
\title{Print method for dp_glm objects}
\usage{
\method{print}{dp_glm}(x, ...)
}
\arguments{
\item{x}{A \code{dp_glm} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print method for dp_glm objects
}
