% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.dp_htest}
\alias{print.dp_htest}
\title{Print method for dp_htest objects}
\usage{
\method{print}{dp_htest}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{dp_htest} object.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments (ignored).}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print method for dp_htest objects
}
