% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_regression.R
\name{print.dp_lm}
\alias{print.dp_lm}
\title{Print method for dp_lm objects}
\usage{
\method{print}{dp_lm}(x, ...)
}
\arguments{
\item{x}{A \code{dp_lm} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print method for dp_lm objects
}
