% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{print.privacy_budget}
\alias{print.privacy_budget}
\title{Print method for privacy_budget objects}
\usage{
\method{print}{privacy_budget}(x, ...)
}
\arguments{
\item{x}{A \code{"privacy_budget"} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print method for privacy_budget objects
}
