% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.summary.dp_lm}
\alias{print.summary.dp_lm}
\title{Print method for summary.dp_lm objects}
\usage{
\method{print}{summary.dp_lm}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{summary.dp_lm} object.}

\item{digits}{Number of digits to print.}

\item{...}{Additional arguments (ignored).}
}
\value{
The object \code{x}, invisibly.
}
\description{
Print method for summary.dp_lm objects
}
